# Api/Educa/EducaApi.Api/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "EducaApi.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/student": {
      "post": {
        "tags": [
          "Student"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StudentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Student"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StudentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/student/get-students/{teacherId}": {
      "get": {
        "tags": [
          "Student"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "teacherId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/student/get-student/{id}": {
      "get": {
        "tags": [
          "Student"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/student/pagination/{teacherId}": {
      "get": {
        "tags": [
          "Student"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Name",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Page",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "OrderByPropety",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SchoolId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "teacherId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/student/{id}": {
      "delete": {
        "tags": [
          "Student"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "StudentDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "observations": {
          "type": "string"
        },
        "class": {
          "type": "string"
        },
        "teacherId": {
          "format": "int32",
          "type": "integer"
        },
        "schoolId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}